module_4/requirements.txt
flask
psycopg2-binary
beautifulsoup4
lxml
pytest
pytest-cov
sphinx
sphinx-rtd-theme
